Move decryption functions to their own module
