+++ lib/ex_uid2/decrypted_token.ex
@@ -1,0 +1,12 @@
+defmodule ExUid2.DecryptedToken do
+  defstruct [
+    :uid,
+    :established,
+    :site_id,
+    :site_key,
+    :identity_scope,
+    :identity_type,
+    :advertising_token_version,
+    :expires
+  ]
+end
